network manager: ignore topology not managed errors
The cluster manager was returning these `ErrNetworkControllerTopologyNotManaged`
errors on the initial sync of the cluster's networks.

Thus, if a customer had a localnet network without subnets provisioned, they
would be locked out of a 4.16 to 4.17 upgrade.

The node network controllers do **not** return these errors (actually, only the
`secondaryNetworkClusterManager` does ...), hence, we can safely swallow and log
the error at the network controller manager layer.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>